Add a failing test case
This file has no .data section and a .bss with zero physical size (but a
non-zero logical size).  It looks like inserting the .bss  section is causing
the problem - `insertAtOffset` walks down the entire list of existing sections
subtracting off section sizes from the offset, but when it reaches the end of
the list, the offset still isn't zero because there is a large gap between the
previous section and the start of the .bss, so it triggers the out of range
error in the Seq.EmptyL case.