Separate ElfSegment and Phdr; fix parsing.
This update adds a mix of changes.

It drops ElfSegment from Phdr, but adds many of the fields directly to
the Phdr record.  This removes the need for the unpopulated
elfSegments.

It introduces special handling for the PT_GNU_STACK segment that
reflects the PT_GNU_STACK is just a way of annotating the program
rather than a reference to file information.

It fixes the parser so that the .bss section will now be identified
as part of the data segment rather than a section after the data
segment.

It adds a warn primitive to the parser so that we have a cleaner
interface to add warnings.