Add `decodeHeaderDynsym` for decoding dynamic symbol tables
Just like `decodeHeaderSymtab` decodes the static function symbol table,
`decodeHeaderDynsym` serves the same role for dynamic function symbol tables.
The functionality of `decodeHeaderDynsym` largely overlaps with the niche that
the `dynamicEntries`/`dynSymEntry` functions provide, so I have included a
comparison to those functions in the Haddocks for `decodeHeaderDynsym`.

This will be useful for eventual fixes for GaloisInc/macaw#277 and
GaloisInc/macaw-loader#12.