Use correct order for lists in buildDynamicMap
We now call `reverse` on the resulting lists after processing the entire buffer
in `buildDynamicMap`, as each call to `insertDynamic` puts new values at the
front of the list, not the end. Calling `reverse` at the end establishes the
same order that entries appear in the original binary, much like how
`gnuLinkedList` calls `reverse` after it has finished.

Fixes #30.