Add elfIsPie to check for position-independent executables
This adds `elfIsPie, a function which returns `True` if the supplied
`ElfHeaderInfo` corresponds to a position-independent executable (PIE). The
logic for `elfIsPie` is adapted from the source code for the `file` binutils
program, which was in turn summarized here:
https://unix.stackexchange.com/a/435038

This adds a `directory` dependency to `elf-edit`, as `directory` is used to
check file permissions as a fallback. This dependency should be cheap, as
`directory` is a GHC boot library.

Fixes #46.